[DEV-103] fix: slot clone & split children by components 컨벤션 반영
